backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pymupdf>=1.23.0
docx2txt>=0.8
python-docx>=1.1.0
google-generativeai>=0.3.0
google-cloud-vision>=3.4.0
bcrypt>=4.1.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.0
pillow>=10.1.0
python-dotenv>=1.0.0